requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.8
greenlet==1.1.2
idna==3.3
psycopg2-binary==2.9.3
python-dateutil==2.8.2
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.30
urllib3==1.26.7
vk-api==11.9.6
